+++ Modules/Module.Settings/Converters/KiloByteConverter.cs
@@ -7,7 +7,7 @@ namespace Module.Settings.Converters
     {
         public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
-            return ((double)(long)value / 1024.00);
+            return (long)value / 1024.00;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
