+++ Services/Service.IO/Protocols/Rx/ProtocolErasePageRx.cs
@@ -8,8 +8,9 @@ namespace Service.IO.Protocols.Rx
     {
         #region ctors
 
-        public ProtocolErasePageRx(byte header, Mode regim)
-            : base(header, 0x5A, regim, new CheckSumXOR7FManager()) { }
+        public ProtocolErasePageRx(byte header, Mode regim) : base(header, 0x5A, regim, new CheckSumXOR7FManager())
+        {
+        }
 
         #endregion
 
@@ -43,7 +44,7 @@ namespace Service.IO.Protocols.Rx
             if (!base.UnPack(buffer))
                 return false;
 
-            PageNumber = (int)(buffer[6] << 21) + (int)(buffer[5] << 14) + (int)(buffer[4] << 7) + buffer[3];
+            PageNumber = (buffer[6] << 21) + (buffer[5] << 14) + (buffer[4] << 7) + buffer[3];
 
             return true;
         }
