+++ Services/Service.IO/Protocols/Rx/ProtocolReadPageRx.cs
@@ -9,8 +9,10 @@ namespace Service.IO.Protocols.Rx
     {
         #region ctors
 
-        public ProtocolReadPageRx(byte header, Mode regim)
-            : base(header, 0x5A, regim, new CheckSumXOR7FManager()) { Line = new PageLine(); }
+        public ProtocolReadPageRx(byte header, Mode regim) : base(header, 0x5A, regim, new CheckSumXOR7FManager())
+        {
+            Line = new PageLine();
+        }
 
         #endregion
 
@@ -47,12 +49,12 @@ namespace Service.IO.Protocols.Rx
                 buffer[24] |= (byte)((Line[k] >> (7 - k)) & (0x01 << k));
                 buffer[25] |= (byte)((Line[7 + k] >> (7 - k)) & (0x01 << k));
             }
-            buffer[24] &= (byte)0x7F;
-            buffer[25] &= (byte)0x7F;
+            buffer[24] &= 0x7F;
+            buffer[25] &= 0x7F;
 
             buffer[26] |= (byte)((Line[14] >> 7) & 0x01);
             buffer[26] |= (byte)((Line[15] >> 6) & 0x02);
-            buffer[26] &= (byte)0x03;
+            buffer[26] &= 0x03;
 
             checkSumManager.Calculate(buffer);
 
@@ -64,9 +66,9 @@ namespace Service.IO.Protocols.Rx
             if (!base.UnPack(buffer))
                 return false;
 
-            PageNumber = (int)(buffer[6] << 21) +
-                (int)(buffer[5] << 14) + (int)(buffer[4] << 7) + buffer[3];
-            PageLineNumber = (int)buffer[7];
+            PageNumber = (buffer[6] << 21) +
+                (buffer[5] << 14) + (buffer[4] << 7) + buffer[3];
+            PageLineNumber = buffer[7];
 
             for (int i = 0; i < PageLine.Size; i++)
                 Line[i] = buffer[8 + i];
